requirements.txt
arabic-reshaper==3.0.0
asgiref==3.4.1
dataclasses==0.8
Django==3.2.22
djangorestframework==3.14.0
djongo==1.3.6
docx==0.2.4
html5lib==1.1
Jinja2==3.0.3
lxml==4.9.3
MarkupSafe==2.0.1
Pillow==8.4.0
PyJWT==2.4.0
pymongo==3.12.3
PyPDF2==3.0.1
python-bidi==0.4.2
python-docx==0.8.11
pytz==2023.3.post1
reportlab==3.6.6
six==1.16.0
sqlparse==0.2.4
typing_extensions==4.1.1
webencodings==0.5.1
xhtml2pdf==0.2.5
